requirements.txt
click == 6.2
motor == 0.5
requests == 2.7.0
SQLAlchemy
tornado == 4.2.1
beautifulsoup4 == 4.5.1
PyJWT == 1.4.2
